upd: reorg/condense fix: typo in comments
