+++ internal/builtins/collector/windows/wmi/collector.go
@@ -19,7 +19,7 @@ import (
 //
 // The individual wmi collector implementations must override Collect and Flush.
 //
-// ID and Inventory are generic and do not need to be overriden unless the
+// ID and Inventory are generic and do not need to be overridden unless the
 // collector implementation requires it.
 
 // Collect metrics
