+++ internal/builtins/collector/windows/wmi/network_interface.go
@@ -56,7 +56,7 @@ type NetInterface struct {
 	exclude *regexp.Regexp
 }
 
-// netInterfaceOptions defines what elements can be overriden in a config file
+// netInterfaceOptions defines what elements can be overridden in a config file
 type netInterfaceOptions struct {
 	ID                   string   `json:"id" toml:"id" yaml:"id"`
 	IncludeRegex         string   `json:"include_regex" toml:"include_regex" yaml:"include_regex"`
