+++ internal/builtins/collector/windows/wmi/network_udp.go
@@ -47,7 +47,7 @@ type NetUDP struct {
 	ipv6Enabled bool
 }
 
-// NetUDPOptions defines what elements can be overriden in a config file
+// NetUDPOptions defines what elements can be overridden in a config file
 type NetUDPOptions struct {
 	ID                   string   `json:"id" toml:"id" yaml:"id"`
 	MetricsEnabled       []string `json:"metrics_enabled" toml:"metrics_enabled" yaml:"metrics_enabled"`
