+++ internal/builtins/collector/windows/wmi/objects.go
@@ -36,7 +36,7 @@ type Objects struct {
 	wmicommon
 }
 
-// objectsOptions defines what elements can be overriden in a config file
+// objectsOptions defines what elements can be overridden in a config file
 type objectsOptions struct {
 	ID                   string   `json:"id" toml:"id" yaml:"id"`
 	MetricsEnabled       []string `json:"metrics_enabled" toml:"metrics_enabled" yaml:"metrics_enabled"`
