+++ internal/builtins/collector/windows/wmi/processor.go
@@ -45,10 +45,10 @@ type Win32_PerfFormattedData_PerfOS_Processor struct {
 type Processor struct {
 	wmicommon
 	numCPU        float64
-	reportAllCPUs bool // may be overriden in config file
+	reportAllCPUs bool // may be overridden in config file
 }
 
-// processorOptions defines what elements can be overriden in a config file
+// processorOptions defines what elements can be overridden in a config file
 type processorOptions struct {
 	ID                   string   `json:"id" toml:"id" yaml:"id"`
 	AllCPU               string   `json:"report_all_cpus" toml:"report_all_cpus" yaml:"report_all_cpus"`
